requirements.txt
mcp>=0.2.0
google-api-python-client>=2.138.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.1